split assets utils
