+++ packages/yoshi/src/commands/start-app.js
@@ -35,7 +35,7 @@ const {
 } = require('../webpack-utils');
 const ServerProcess = require('../server-process');
 const detect = require('detect-port');
-const { watchPublicFolder } = require('./utils/assets');
+const { watchPublicFolder } = require('./utils/copy-assets');
 const rootApp = require('yoshi-config/root-app');
 
 const host = '0.0.0.0';
