+++ packages/yoshi/src/commands/utils/build-libs.js
@@ -4,7 +4,7 @@ const execa = require('execa');
 const chalk = require('chalk');
 const globby = require('globby');
 
-async function buildLibs(libs) {
+module.exports = async function buildLibs(libs) {
   // Clean tmp folders
   await Promise.all(
     libs.map(app => {
@@ -38,6 +38,4 @@ async function buildLibs(libs) {
 
     process.exit(1);
   }
-}
-
-module.exports = buildLibs;
+};
